requirements.txt
boto3==1.26.115
botocore==1.29.130
Jinja2==3.1.2
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0
PyYAML==6.0.1
redshift_connector==2.0.910
setuptools==68.0.0
mkdocs
mkdocs-material=="9.*"